Moving more towards cleaner but ordered XML child handler

+++ plugins/engine/GEK/Engine/Resources.h
@@ -11,7 +11,7 @@
 
 namespace Gek
 {
-    template <typename TYPE, int UNIQUE>
+    template <typename TYPE, uint8_t UNIQUE>
     struct Handle
     {
         TYPE identifier;
